+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 -   Add `kwargs` to `BashExecutionParameters` [#90](https://github.com/getindata/dbt-airflow-factory/issues/90)
+-   Correcting required packages [#97](https://github.com/getindata/dbt-airflow-factory/issues/97)
 
 ## [0.32.0] - 2023-07-04