+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## [Unreleased]
 
+## [0.34.0] - 2023-08-10
+
+-   Add `MS Teams` notifications handler
+
 ## [0.33.0] - 2023-08-04
+
 -   Add `kwargs` to `BashExecutionParameters` [#90](https://github.com/getindata/dbt-airflow-factory/issues/90)
 -   Correcting required packages [#97](https://github.com/getindata/dbt-airflow-factory/issues/97)
 
@@ -150,7 +155,9 @@ This version brings compatibility with `dbt 1.0`.
 
 -   Initial implementation of `dbt_airflow_manifest_parser` library.
 
-[Unreleased]: https://github.com/getindata/dbt-airflow-factory/compare/0.33.0...HEAD
+[Unreleased]: https://github.com/getindata/dbt-airflow-factory/compare/0.34.0...HEAD
+
+[0.34.0]: https://github.com/getindata/dbt-airflow-factory/compare/0.33.0...0.34.0
 
 [0.33.0]: https://github.com/getindata/dbt-airflow-factory/compare/0.32.0...0.33.0